+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Changelog for AgentMap v1.0.2
+# Changelog for AgentMap v1.0
 
 ## v1.0.0
 
@@ -10,6 +10,9 @@ Complete rewrite.
 
   * [AgentMap] `get_prop(am, :size)` is optimized. From now it may temporary
     behave a little unaccurate upwards (in some rare cases).
+  * [AgentMap] new `tiny: true` option for `get_and_update4`, `update/4`,
+    `update!/4`, `cast/4` and `put_new_lazy/4`;
+  * [AgentMap] new `upd_prop/3` call;
   * [AgentMap.Multi] `get_and_update/4`, `update/4`, `cast/4` are now have a
     fixed priority `{:avg, +1}`